chore(utils): move datetime utils to separate module
